Update to PlayN 1.9.1
The project could not be built using 1.9 since there was a dependency
on read-only files on Google Code, which no longer exists. Michael Bayne
was kind enough to update PlayN to 1.9.1 to remove that dependency.
This required separating playn.version and tripleplay.version, but now
mvn package works as expected.